测试 检测 出来 声音

+++ managed_components/lvgl__lvgl/tests/main.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python3
 
 import argparse
-import errno
 import shutil
 import subprocess
 import sys
@@ -17,6 +16,7 @@ sys.path.append(lvgl_script_path)
 wayland_dir = os.path.join(lvgl_test_dir, "wayland_protocols")
 wayland_protocols_dir = os.path.realpath("/usr/share/wayland-protocols")
 
+from perf import perf_test_options
 from LVGLImage import LVGLImage, ColorFormat, CompressMethod
 
 # Key values must match variable names in CMakeLists.txt.
@@ -25,7 +25,6 @@ build_only_options = {
     'OPTIONS_16BIT': 'Minimal config, 16 bit color depth',
     'OPTIONS_24BIT': 'Normal config, 24 bit color depth',
     'OPTIONS_FULL_32BIT': 'Full config, 32 bit color depth',
-    'OPTIONS_VG_LITE': 'VG-Lite simulator with full config, 32 bit color depth',
 }
 
 if platform.system() != 'Windows':
@@ -35,6 +34,7 @@ test_options = {
     'OPTIONS_TEST_SYSHEAP': 'Test config, system heap, 32 bit color depth',
     'OPTIONS_TEST_DEFHEAP': 'Test config, LVGL heap, 32 bit color depth',
     'OPTIONS_TEST_VG_LITE': 'VG-Lite simulator with full config, 32 bit color depth',
+    'OPTIONS_TEST_RISCV_V': 'RISC-V Vector emulation with full config, 32 bit color depth',
 }
 
 
@@ -176,13 +176,13 @@ def generate_code_coverage_report():
     os.chdir(lvgl_test_dir)
     delete_dir_ignore_missing('report')
     os.mkdir('report')
-    root_dir = os.pardir
+    root_dir = os.path.dirname(lvgl_test_dir)  # Get parent directory of tests (lvgl directory)
     html_report_file = 'report/index.html'
-    cmd = ['gcovr', '--gcov-ignore-parse-errors', 'negative_hits.warn', 
+    cmd = ['gcovr', '--gcov-ignore-parse-errors', 
            '--root', root_dir, '--html-details', '--output',
            html_report_file, '--xml', 'report/coverage.xml',
-           '-j', str(os.cpu_count()), '--print-summary',
-           '--html-title', 'LVGL Test Coverage', '--filter', r'../src/.*/lv_.*\.c']
+           '-j', str(os.cpu_count()), '--print-summary', '--merge-mode-functions=merge-use-line-min',
+           '--html-title', 'LVGL Test Coverage', '--filter', os.path.join(root_dir, 'src/.*/lv_.*\\.c')]
 
     subprocess.check_call(cmd)
     print("Done: See %s" % html_report_file, flush=True)
@@ -213,6 +213,19 @@ def generate_test_images():
                     print(f"converting {os.path.basename(png)}, format: {fmt.name}, compress: {compress_name}")
 
 
+def clean_build_dirs_with_filter(build_dir, clean_filters):
+    for entry in os.listdir(build_dir):
+        entry_path = os.path.join(build_dir, entry)
+
+        if any(entry.endswith(filter) for filter in clean_filters):
+            continue
+
+        if os.path.isfile(entry_path):
+            os.remove(entry_path)
+        elif os.path.isdir(entry_path):
+            shutil.rmtree(entry_path)
+
+
 if __name__ == "__main__":
     epilog = '''This program builds and optionally runs the LVGL test programs.
     There are two types of LVGL tests: "build", and "test". The build-only
@@ -223,7 +236,7 @@ if __name__ == "__main__":
     parser = argparse.ArgumentParser(
         description='Build and/or run LVGL tests.', epilog=epilog)
     parser.add_argument('--build-options', nargs=1,
-                        choices=list(chain(build_only_options, test_options)),
+                        choices=list(chain(build_only_options, test_options, perf_test_options)),
                         help='''the build option name to build or run. When
                         omitted all build configurations are used.
                         ''')
@@ -239,6 +252,8 @@ if __name__ == "__main__":
                         help='Update test image using LVGLImage.py script')
     parser.add_argument('--auto-clean', action='store_true', default=False,
                         help='Automatically clean build directories')
+    parser.add_argument('--keep-report', action='store_true', default=False,
+                        help='Skip cleaning gcov report files when --auto-clean and --report are enabled')
 
     args = parser.parse_args()
 
@@ -256,8 +271,18 @@ if __name__ == "__main__":
         else:
             options_to_build = test_options
 
+    clean_build_dirs = []
     for options_name in options_to_build:
         is_test = options_name in test_options
+        is_perf_test = options_name in perf_test_options
+        if is_perf_test:
+            perf_test_script = os.path.join(lvgl_test_dir, "perf.py")
+            try:
+                subprocess.check_call([perf_test_script, *(sys.argv[1:])])
+            except subprocess.CalledProcessError as e:
+                sys.exit(e.returncode)
+            continue
+
         build_type = 'Debug'
         build_tests(options_name, build_type, args.clean)
         if is_test:
@@ -265,10 +290,31 @@ if __name__ == "__main__":
                 run_tests(options_name, args.test_suite)
             except subprocess.CalledProcessError as e:
                 sys.exit(e.returncode)
+
         if args.auto_clean:
             build_dir = get_build_dir(options_name)
-            print("Removing " + build_dir)
-            shutil. rmtree(build_dir)
+
+            if args.report:
+                # Keep the files that gcovr analysis depends on and delete
+                # the rest to solve the storage capacity limit of GitHub CI
+                clean_filters = ['CMakeFiles', '.c']
+                clean_build_dirs_with_filter(build_dir, clean_filters)
+
+                if args.keep_report:
+                    # Retain the gcov report files for subsequent automated coverage analysis.
+                    print(f"Keeping {build_dir} for report")
+                else:
+                    print(f"Append {build_dir} to clean list")
+                    clean_build_dirs.append(build_dir)
+            else:
+                # Remove all build directories directly
+                print(f"Removing {build_dir}")
+                shutil.rmtree(build_dir)
 
     if args.report:
         generate_code_coverage_report()
+
+    # Clean all build directories after report
+    for build_dir in clean_build_dirs:
+        print(f"Removing {build_dir}")
+        shutil.rmtree(build_dir)
